Update Configuration, CSV Output
After a lot of work, I decided to solve the problem of sharing
configuration between DA and the caller by abandoning that idea,
altogether.  Instead of having a configuration that is set and read
separately, as a unit for all of DA, each method exposes all available
configuration through its signature.  Not all methods have been
converted and some configuration still remains.  But, this should now be
removed and future development will not expand the old configuration
approach.

DA now allows output in CSV format.

DV now calls chop() with CSV as the default output.