DataView: WIP: Read Window from Figure Pan/Zoom
This commit actually sets up the callback from any draw event.  It also
harvests the new x-axis range values, which can be used to calculate
new window values.  But, it's tricky, and there is also the problem of
extraneous events.  If the figure is updated, externally, using sliders,
this will also cause this event, so I need to figure out how to filter
those as it would be an unecessary update and, possibly, a loop.