Active, View-Based Translation Implemented
This completes the inital transformations I wanted to implement.  Its
hardly complete.  This patch added the active view- or tag-based
translation transformation code that we wanted for a while.  It's pretty
well tested as I was able to visualize retrograde motion using it on
some of the "astro" data we took a couple months ago.