Refactor DVdialog Class into DVdialogHelper
DVdialog class was a typical tkinter class for building dialog boxes.
But, my use of it was incorrect, calling it with launchDialog and
lambda, which did not store the values for later.  This can be
accomplished with functools.partial.  But, I did this instead, and it
works, so I'm going to leave it.